WIP Handle logical attnum
A new attnum column is added to pg_attribute, to store the logical column
position.  The default is to have the same as the physical position.

A relation is now expanded in logical order, for instance:

- A plain SELECT * FROM table
- A RETURNING * clause in a DML command
- A COPY table FROM/TO without explicit column names

Note that there's for now no possibility to declare a different logical
position.

psql displays a table columns information using the logical order rather the
physical order, and if verbose emits an addition "Physical order" footer if the
logical layout is different from the physical one.

Many problems that still exist, for instance deparsing view that didn't have
explicit column list and later requires it (e.g. after renaming attributes).

Author: Julien Rouhaud
Reviewed-by: FIXME
Discussion: FIXME