POC: Add infrastructure to specify catalog attributes logical position.
This introduces a new optional macro ATTNUM(attnum) which added to specify the
logical order of the attributes declared in a struct, if the regular struct
order isn't wanted.

Catalog.pm and genbki.pl will use that information to generate the data in the
correct oder.

Sanity check are performed to make sure that the declared ATTNUMs are a
gapeless sequence from 1 to the maximum number of attributes.

As an example, move relname near the end of pg_class struct and attname near
the end of pg_attribute.

Author: Julien Rouhaud
Discussion: FIXME